refactor: Move ObjectOrErrorMessage to own file
This is a step on the way to generalising FilterOrErrorMessage for things
other than filters.
